+++ springrain/src/main/java/org/springrain/frame/util/Page.java
@@ -21,7 +21,7 @@ public class Page implements Serializable{
 	/** 每页多少行 */
 	private int pageSize=20;
 	/** 数据总行数 */
-	private long totalCount = 0;
+	private int totalCount = 0;
 	/** 总共可以分多少页 */
 	private int pageCount;
 	/**排序方式 desc asc*/
@@ -133,9 +133,9 @@ public class Page implements Serializable{
 	 * 
 	 * @param totalCount
 	 */
-	public void setTotalCount(long totalCount) {
+	public void setTotalCount(int totalCount) {
 		this.totalCount = totalCount;
-		this.pageCount = (int) ((totalCount+pageSize-1) / pageSize);
+		this.pageCount = (totalCount+pageSize-1) / pageSize;
 				
 		if(pageIndex>=pageCount){
 			this.lastPage=true;
